requirements.txt
streamlit==1.44.0

google.generativeai== 0.8.4
sqlite3==3.45.3